chore: move utils/session to utils/space
